plot: move show/diff to API
